webscraper/webscraper/requirements.txt
flask==2.2.5
flask-cors==5.0.1
google-generativeai==0.8.4
python-docx==1.1.0
docx2pdf==0.1.8
werkzeug==3.1.3
